Remove unwrapArray
This caused API of page.onResourceRequested to change with phantom
1.9.0 (first argument was `request`, it became an array [`request`,
`networkRequest`]).

Now use fn.apply to pass arguments as they are received. This ensure
that any change in phantomjs is transparently reflected without
interference from node-phantom.

NOTE: This change the API of onError and onPrompt which where already
affected by this bug. They now get the same parameters as described on
https://github.com/ariya/phantomjs/wiki/API-Reference-WebPage.